sliding text - refactored common en formatting functions into common module
