+++ apps.json
@@ -226,10 +226,11 @@
     "storage": [
       {"name":"slidingtext.app.js","url":"slidingtext.js"},
       {"name":"slidingtext.img","url":"slidingtext-icon.js","evaluate":true},
-      {"name":"slidingtext.local.en.js","url":"slidingtext.local.en.js"},
-      {"name":"slidingtext.local.en2.js","url":"slidingtext.local.en2.js"},
-      {"name":"slidingtext.local.fr.js","url":"slidingtext.local.fr.js"},
-      {"name":"slidingtext.local.jp.js","url":"slidingtext.local.jp.js"},
+      {"name":"slidingtext.locale.en.js","url":"slidingtext.locale.en.js"},
+      {"name":"slidingtext.locale.en2.js","url":"slidingtext.locale.en2.js"},
+      {"name":"slidingtext.utils.en.js","url":"slidingtext.utils.en.js"},
+      {"name":"slidingtext.locale.fr.js","url":"slidingtext.locale.fr.js"},
+      {"name":"slidingtext.locale.jp.js","url":"slidingtext.locale.jp.js"},
       {"name":"slidingtext.dtfmt.js","url":"slidingtext.dtfmt.js"}
     ]
   },
