+++ apps/slidingtext/slidingtext.locale.en2.js
@@ -1,42 +1,12 @@
 var DateFormatter = require("slidingtext.dtfmt.js");
-
-const numberStr = ["ZERO","ONE", "TWO", "THREE", "FOUR", "FIVE",
-    "SIX", "SEVEN","EIGHT", "NINE", "TEN",
-    "ELEVEN", "TWELVE", "THIRTEEN", "FOURTEEN",
-    "FIFTEEN", "SIXTEEN", "SEVENTEEN", "EIGHTEEN",
-    "NINETEEN", "TWENTY"];
-const tensStr = ["ZERO", "TEN", "TWENTY", "THIRTY", "FOURTY",
-    "FIFTY"];
-
-function hoursToText(hours){
-    hours = hours % 12;
-    if(hours == 0){
-        hours = 12;
-    }
-    return numberStr[hours];
-}
-
-function numberToText(value){
-    var word1 = '';
-    var word2 = '';
-    if(value > 20){
-        var tens = (value / 10 | 0);
-        word1 = tensStr[tens];
-        var remainder = value - tens * 10;
-        if(remainder > 0){
-            word2 = numberStr[remainder];
-        }
-    } else if(value > 0) {
-        word1 = numberStr[value];
-    }
-    return [word1,word2];
-}
+const hoursToText = require("slidingtext.utils.en.js").hoursToText;
+const numberToText = require("slidingtext.utils.en.js").numberToText;
 
 class EnglishTraditionalDateFormatter extends DateFormatter {
     constructor() {
         super();
     }
-    name(){return "English Traditional";}
+    name(){return "English (Traditional)";}
     formatDate(date){
         var mins = date.getMinutes();
         var hourOfDay = date.getHours();
@@ -56,17 +26,26 @@ class EnglishTraditionalDateFormatter extends DateFormatter {
         }
         var mins_txt;
         var from_to;
+        var mins_value;
         if(mins > 30){
+            mins_value = 60-mins;
             from_to = "TO";
-            mins_txt = numberToText(60-mins);
+            mins_txt = numberToText(mins_value);
         } else {
+            mins_value = mins;
             from_to = "PAST";
-            mins_txt = numberToText(mins);
+            mins_txt = numberToText(mins_value);
         }
         if(mins_txt[1] != '') {
             return ['', mins_txt[0], mins_txt[1], from_to, hours];
         } else {
-            return ['', mins_txt[0], from_to, '', hours];
+            if(mins_value % 5 == 0) {
+                return ['', mins_txt[0], from_to, '', hours];
+            } else if(mins_value == 1){
+                return ['', mins_txt[0], 'MINUTE', from_to, hours];
+            } else {
+                return ['', mins_txt[0], 'MINUTES', from_to, hours];
+            }
         }
     }
 }
