+++ apps/slidingtext/slidingtext.locale.jp.js
